Fix integration tests (#383)
* Re-enable bot integration tests

* Connect to a MongoDB database to pass the tests

The commit [1] caused the failure of tests because line 191 of
the start.js file in that commit:
`
const config = await Config.findOne({ maintenance: true });
`
is hanging and doesn't respond, causing a timeout and failure
of the tests. Based on [2] the reason is that we need to call
mongoose.connect(...) to connect to the database before that
line. That's why I called `connect` function from
`db_connect.js` file in the `start.js` file. There are other
similar lines in the `start.js` file like:
`
const order = await Order.findOne({ _id: orderId });
`
That doesn't cause timeout and failure of the CI and I don't
know the reason.

Fixes https://github.com/lnp2pBot/bot/issues/378

[1] https://github.com/lnp2pBot/bot/commit/ed409030deb594f689413550ce769205b36669ab
[2] https://stackoverflow.com/questions/65408618/mongooseerror-operation-users-findone-buffering-timed-out-after-10000ms

Co-authored-by: Parham <parhaamsaremi@gmail.com>

---------

Co-authored-by: Parham <parhaamsaremi@gmail.com>