CI: downgrade to v18
The change done in [1] is good because 14.x version was
too old, but 20.x is too new. The new Ubunutu LTS, released
last month, doesn't even have such a bleeding edge Node
version. Let's downgrade to 18.x which is a good sweet spot:
new enough, but also stable enough.

This way we make sure LNp2pBot is compatible with the same
version of NodeJS shipped by the last Ubuntu LTS.

[1] 439d043b1d0ccee9d1f29c3d3d4513b3fd9ac933