CI: upgrade Ubuntu to 24.04
GitHubActions' runner image for 'ubuntu-latest' is still running
old LTS 22.04, but we can instead use the 'container:' attribute
to use a docker image in order to use the new Ubuntu LTS 24.04.

This is the most sensible move especially considering that we
have changed CI recently to run on new NodeJS 18.x by default
(Ubuntu 22.04's default NodeJS version is much older than v18).

NOTE: MongoDB service start had to be reworked in this commit to
make it still work in this new distro version.