Convert JS to TS (commits 7-13 of PR435) (#557)
* util/index: convert js to ts

The languageModel & fiatModel interface created to type-check
the data and ensure that it conforms to
the expected structure and format.
This can help you avoid errors and bugs when
working with the data in your code.
This commit enables resolveJsonModule to import module
with '.json' extension.

Co-authored-by: webwarrior <reg@webwarrior.ws>

* jobs: convert js to ts

The --downlevelIteration flag is a TypeScript compiler
option that enables support for iterating over new
concepts like Map, Set, or Generator in older JavaScript
runtimes. By default, TypeScript targets ES3, which does
not support these features. If you use a for...of loop or
a spread operator on an iterable object, you may get an error.
Use Date instead of Date.toISOString cause paid_at has type
Date and during the conversion from js to ts,
we got compilation errors.

Co-authored-by: webwarrior <reg@webwarrior.ws>

* bot/start: fixing types & import/exports

Co-authored-by: webwarrior <reg@webwarrior.ws>

* bot/validation: convert js to ts

Using null instead of a boolean/undefined type is better.

* lnurl/lnurl-pay: convert js to ts

I had to change the '|', otherwise typescript would
complain this error msg:
```
The '|' operator is not allowed for boolean types.
Consider using '||' instead.
```

Co-authored-by: webwarrior <reg@webwarrior.ws>

* refactor: correcting a bad practice

* refactor: there's no need for isInt()

---------

Co-authored-by: Mehrshad <code.rezaei@gmail.com>