refactor: removes 'orders' field from the ICommunity interface
The 'orders' field is not part of the Mongoose document, but rather it's something that's calculated on-the-fly
as part of the response to the `/findcomms` command for each community. Converting the mongoose document to a
plain object gets rid of the TSC errors without having to introduce an 'orders' field to the ICommunity interface.

This solution is preferred as the ICommunity interface should ideally mirror what's stored in the database.