CI: check formatting at the end of the pipeline (#691)
* package(-lock).json: upgrade to prettier 3.6.2

* *.ts: apply prettier formatting

Based on this comment:
https://github.com/lnp2pBot/bot/pull/618#issuecomment-2609866298

* GitHubCI: run prettier at the end

To make sure that all contributions are according to the style.
This also modifies CONTRIBUTING.md to mention CI wrt PRs.

* CI: fix git error

Workaround suggested in:
https://github.com/actions/checkout/issues/363

Error was (when running 'git diff'):
```
warning: Not a git repository. Use --no-index to compare two paths outside a working tree
```

* CI: simplify build to avoid git diff noise

* Create a .git-blame-ignore-revs file

This file contains a list of git hashes of revisions to be ignored by git.
These revisions are considered "unimportant" in that they are unlikely to
be what you are interested in when using the git-blame command (or the
blame view in Github); like when formatting with Prettier.